requirements.txt
easyocr==1.6.2
ipython==8.10.0
matplotlib==3.7.0
numpy==1.24.1
opencv_python_headless==4.5.4.60
Pillow==9.4.0
streamlit==1.18.1
